Split ToolRegistry into single-responsibility collaborators
ToolRegistry had grown to 397 code lines doing four jobs at once: holding the
catalogue of 32 tools, describing their JSON schemas, carrying out their
handlers, and dispatching with policy enforcement. Nine of those tools were mine
from this branch, so I contributed to the problem.

Split along those seams, with no change to the public surface
(ToolRegistry.tools/find/list/execute/capabilities/symbolize all behave as
before, and all 32 tools remain registered):

  Tool           value object, with a write? predicate replacing the
                 risk.end_with?("write") check that execute did inline
  ToolSchemas    JSON Schema fragments — pure data
  ToolHandlers   the callables, one per tool
  ToolCatalogue  which tools exist, at what scope and risk
  ToolRegistry   lookup, policy enforcement, dispatch
  KeyCoercion    recursive symbolize/stringify, previously buried in the registry

ToolRegistry drops from 397 to 44 code lines and no longer needs its
Metrics/ModuleLength suppression.

Note on measurement: models/order.rb looked like the worst offender at 712 lines
but is only 246 lines of code to 466 of YARD documentation, so it was left alone
— refactoring it would have been churn against a well-documented class rather
than a design fix.

KeyCoercion is called explicitly rather than mixed in. module_function makes the
copies private when a module is extended, which silently broke
ToolRegistry.symbolize for Agent::OrderPreview; the specs caught it.

1,039 examples, 0 failures; RuboCop clean.
